renamed "application" to "task", so as not to be confused with "application" == finished program
